+++ packages/terra-application-name/src/terra-dev-site/doc/example/ApplicationHeaderNameStandard.jsx
@@ -1,13 +1,15 @@
 import React from 'react';
 import ApplicationHeaderLayout from 'terra-application-header-layout';
 import Image from 'terra-image';
+import classNames from 'classnames/bind';
 /* eslint-disable import/no-extraneous-dependencies, import/no-unresolved, import/extensions */
 import Placeholder from 'terra-application-name/lib/terra-dev-site/doc/common/Placeholder';
 import ApplicationHeaderName from 'terra-application-name/lib/ApplicationHeaderName';
 /* eslint-enable import/no-extraneous-dependencies, import/no-unresolved, import/extensions */
+import demoColors from '../../test/application-name/demoStyles.scss';
 
+const cx = classNames.bind(demoColors);
 const demosize = { height: '50px', width: '100%' };
-const demobgcolor = { backgroundColor: 'rgb(140, 161, 177)' };
 const ApplicationHeaderNameStandard = () => (
   <ApplicationHeaderLayout
     style={demosize}
@@ -15,7 +17,7 @@ const ApplicationHeaderNameStandard = () => (
       <ApplicationHeaderName
         title="App-Name"
         accessory={<Image alt="Terra Logo" variant="rounded" src="https://github.com/cerner/terra-framework/raw/master/terra.png" height="26px" width="26px" isFluid />}
-        style={demobgcolor}
+        className={cx(['demo-background-color'])}
       />
     )}
     extensions={<Placeholder text="Extensions" width="100px" />}