+++ packages/terra-infinite-list/CHANGELOG.md
@@ -4,6 +4,11 @@ Changelog
 Unreleased
 ----------
 
+3.3.0 - (April 9, 2019)
+------------------
+### Changed
+* Changed documentation placeholder example colors to meet color contrast ratio
+
 3.2.0 - (April 2, 2019)
 ------------------
 ### Added